Move json support related classes to common package and make them package-private.
