+++ rsql-jpa/src/main/java/io/github/perplexhub/rsql/JsonbConfiguration.java
@@ -1,4 +1,4 @@
-package io.github.perplexhub.rsql.jsonb;
+package io.github.perplexhub.rsql;
 
 import lombok.Builder;
