+++ rsql-jpa/src/main/java/io/github/perplexhub/rsql/SortUtils.java
@@ -7,7 +7,6 @@ import java.util.regex.Pattern;
 import java.util.stream.Collectors;
 import java.util.stream.Stream;
 
-import io.github.perplexhub.rsql.jsonb.JsonbSupport;
 import jakarta.persistence.criteria.CriteriaBuilder;
 import jakarta.persistence.criteria.Expression;
 import jakarta.persistence.criteria.Order;
